Updated db to not reseed users on load.

+++ myrate/myrate_backend/server/db/conn.js
@@ -26,35 +26,41 @@ module.exports = {
       const collection = client.db("media").collection("users");
       // The drop() command destroys all data from a collection.
         // Make sure you run it against proper database and collection.
-        try {
-          collection.drop();
-
-        }
-        catch (err)
+        if(collection.countDocuments() == 0)
         {
-
+          try {
+            collection.drop();
+  
+          }
+          catch (err)
+          {
+  
+          }
+  
+          // make a bunch of time series data
+          let userData = [];
+  
+          for (let i = 0; i < 5000; i++) {
+              const firstName = faker.name.firstName();
+              const lastName = faker.name.lastName();
+              let newUser = {
+                  firstName,
+                  lastName,
+                  day_joined: faker.date.past(),
+                  email: faker.internet.email(firstName, lastName),
+                  username: faker.name.firstName() + faker.random.word() + faker.random.alphaNumeric(),
+                  password: faker.internet.password(),
+              };
+  
+              userData.push(newUser);
+          }
+          collection.insertMany(userData);
+  
+          console.log("User Database seeded");
         }
-
-        // make a bunch of time series data
-        let userData = [];
-
-        for (let i = 0; i < 5000; i++) {
-            const firstName = faker.name.firstName();
-            const lastName = faker.name.lastName();
-            let newUser = {
-                firstName,
-                lastName,
-                day_joined: faker.date.past(),
-                email: faker.internet.email(firstName, lastName),
-                username: faker.name.firstName() + faker.random.word() + faker.random.alphaNumeric(),
-                password: faker.internet.password(),
-            };
-
-            userData.push(newUser);
+        else {
+          console.log("Users are already in database. Did not seed.");
         }
-        collection.insertMany(userData);
-
-        console.log("User Database seeded");
 
 
       return callback(err);
